Update passwords.py
This Python program utilizes TKinter to create a user interface in which the user can create a password. The program catches passwords that don't meet requirements and suggests the user try again. After a successful password is made, the program disables the reset and submit buttons and allows the user to exit the window.